Most stable release yet!
The board now gets disabled for clicking whil the AI is making a move.  The game can be ended on the Human or AI player's turn.  Just one small issue remains.  The AI is still able to check his own king by moving other pieces, or by moving the king.